Move sign_incrementally to src/signer/incremental.rs as well.
